chat/requirements.txt
flask==2.3.3
flask-cors==4.0.0
transformers>=4.40.0
tokenizers>=0.15.0
torch>=2.5.0
numpy>=1.24.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests>=2.31.0
